refactor: 파일 경로 문제 수정을 위한 추가 수정
